Refactor `SemestersEnum`
- remove `@Disabled` tags for tests
