+++ src/main/java/nus/climods/logic/parser/parameters/SemesterTypeParameter.java
@@ -1,13 +1,13 @@
 package nus.climods.logic.parser.parameters;
 
-import org.openapitools.client.model.ModuleCondensed;
+import org.openapitools.client.model.SemestersEnum;
 
 import nus.climods.logic.parser.ParserUtil;
 
 /**
  * Represents a Semester Type parameter
  */
-public class SemesterTypeParameter extends OptionalPositionalParameter<ModuleCondensed.SemestersEnum> {
+public class SemesterTypeParameter extends OptionalPositionalParameter<SemestersEnum> {
     public static final String INVALID_INPUT_MESSAGE = "Semester should not be empty!";
     public static final String PARSE_EXCEPTION_MESSAGE = "Invalid Semester: %s";
     private static final int LESSON_TYPE_INDEX = 1;
