+++ src/test/java/nus/climods/storage/module/JsonModuleSummaryListStorageTest.java
@@ -10,7 +10,6 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.Optional;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.openapitools.client.api.ModulesApi;
@@ -54,8 +53,6 @@ class JsonModuleSummaryListStorageTest {
             "notJsonFormatModuleSummaryList.json"));
     }
 
-    //TODO: Not to sure why this test case failed, temporarily ignoring it for now.
-    @Disabled
     @Test
     public void readValidJsonFormat_success() throws Exception {
         assertTrue(readModuleSummaryList("validJsonFormatModuleSummaryList.json").isPresent());
@@ -67,8 +64,6 @@ class JsonModuleSummaryListStorageTest {
      *
      * @throws Exception any exception that may occur.
      */
-    //TODO: Not to sure why this test case failed, temporarily ignoring it for now.
-    @Disabled
     @Test
     public void saveAndReadModuleSummaryList_success() throws Exception {
         Path filePath = testFolder.resolve("TempModuleSummaryList.json");
