+++ src/test/java/org/openapitools/client/api/ModulesApiTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.openapitools.client.ApiException;
 import org.openapitools.client.model.Module;
@@ -27,8 +26,6 @@ public class ModulesApiTest {
         assertTrue(response.size() > 0);
     }
 
-    //TODO: Not to sure why this test case failed, temporarily ignoring it for now.
-    @Disabled
     @Test
     public void acadYearModuleListJsonGetTest() throws ApiException {
         String acadYear = "2022-2023";
